Revamped the coordinate system and added live preview
The user can now use boxes to center the text properly on the image. These boxes can be made to perfectly enclose a blank space in certificate. User also gets a live preview of what their selection will look like on the final certificate. In addition to this the application now crops down the image for ease of use.